create base fragment for make code more understanding and remove duplicate code add BackPressed interface
