+++ .idea/misc.xml
@@ -4,6 +4,7 @@
     <option name="filePathToZoomLevelMap">
       <map>
         <entry key="..\:/android/PopularLibraries/app/src/main/res/layout/activity_main.xml" value="0.15625" />
+        <entry key="..\:/android/PopularLibraries/app/src/main/res/layout/fragment_user_info.xml" value="0.23854166666666668" />
         <entry key="..\:/android/PopularLibraries/app/src/main/res/layout/fragment_users.xml" value="0.20416666666666666" />
         <entry key="..\:/android/PopularLibraries/app/src/main/res/layout/item_users.xml" value="0.20416666666666666" />
       </map>
