+++ app/src/main/java/com/example/popularlibraries/ui/base/BackButtonListener.kt
@@ -1,0 +1,5 @@
+package com.example.popularlibraries.ui.base
+
+interface BackButtonListener {
+    fun backPressed(): Boolean
+}
