+++ app/src/main/java/com/example/popularlibraries/ui/base/BaseFragment.kt
@@ -1,0 +1,31 @@
+package com.example.popularlibraries.ui.base
+
+import android.os.Bundle
+import android.view.LayoutInflater
+import android.view.View
+import android.view.ViewGroup
+import androidx.viewbinding.ViewBinding
+import moxy.MvpAppCompatFragment
+
+typealias Inflate<T> = (LayoutInflater, ViewGroup?, Boolean) -> T
+abstract class BaseFragment<VB: ViewBinding> (
+    private val inflate : Inflate<VB>
+        ): MvpAppCompatFragment(), BackButtonListener {
+    private var _binding : VB? = null
+    val binding get() = _binding!!
+
+    override fun onCreateView(
+        inflater: LayoutInflater,
+        container: ViewGroup?,
+        savedInstanceState: Bundle?
+    ): View {
+        super.onCreateView(inflater, container, savedInstanceState)
+        _binding = inflate.invoke(inflater, container, false)
+        return binding.root
+    }
+
+    override fun onDestroyView() {
+        super.onDestroyView()
+        _binding = null
+    }
+}
