+++ app/src/main/java/com/example/popularlibraries/ui/users/view/UsersFragment.kt
@@ -1,21 +1,16 @@
 package com.example.popularlibraries.ui.users.view
 
 import android.os.Bundle
-import android.view.LayoutInflater
 import android.view.View
-import android.view.ViewGroup
 import androidx.recyclerview.widget.LinearLayoutManager
 import com.example.popularlibraries.App
 import com.example.popularlibraries.databinding.FragmentUsersBinding
 import com.example.popularlibraries.model.UsersModelImplementation
+import com.example.popularlibraries.ui.base.BaseFragment
 import com.example.popularlibraries.ui.users.presenter.UsersPresenter
-import moxy.MvpAppCompatFragment
 import moxy.ktx.moxyPresenter
 
-class UsersFragment : MvpAppCompatFragment(), UsersView {
-
-    private var _binding: FragmentUsersBinding? = null
-    private val binding get() = _binding!!
+class UsersFragment : BaseFragment<FragmentUsersBinding>(FragmentUsersBinding::inflate), UsersView {
 
     private val presenter by moxyPresenter {
         UsersPresenter(App.instance.router, UsersModelImplementation())
@@ -25,19 +20,15 @@ class UsersFragment : MvpAppCompatFragment(), UsersView {
         UserAdapter(presenter.usersListPresenter)
     }
 
-    override fun onCreateView(
-        inflater: LayoutInflater,
-        container: ViewGroup?,
-        savedInstanceState: Bundle?
-    ): View {
-        super.onCreateView(inflater, container, savedInstanceState)
-        _binding = FragmentUsersBinding.inflate(inflater)
-        return binding.root
-    }
 
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
         super.onViewCreated(view, savedInstanceState)
         initView()
+    }
+
+    override fun backPressed(): Boolean {
+        presenter.backPressed()
+        return true
     }
 
     private fun initView() {
@@ -47,11 +38,6 @@ class UsersFragment : MvpAppCompatFragment(), UsersView {
             usersList.adapter = adapter
         }
 
-    }
-
-    override fun onDestroyView() {
-        super.onDestroyView()
-        _binding = null
     }
 
     override fun updateUsers() {
